Feature: 完善矩阵库基础能力以适配 iLQR
1. 重构 matrix 核心类型和姿态相关类型，整理矩阵、向量、切片、四元数、欧拉角等实现。
2. 新增 CholeskyDecomposition，并更新最小二乘、伪逆等线性代数工具。
3. 增加 NuttX initializer_list 支持，满足矩阵库初始化列表使用需求。
4. 更新 iLQR 子模块指针，引入切片赋值、Matrix 聚合头精简和 Cholesky 解耦改动。
5. 新增 clang-format 配置并同步调整 Bezier 相关代码格式。

+++ src/lib/bezier/BezierN.cpp
@@ -47,7 +47,7 @@ namespace
  * Generic in-place bezier implementation. Leaves result in first element.
  *
  */
-template <typename Scalar, size_t D>
+template <typename Scalar, int D>
 void calculateBezier(matrix::Vector<Scalar, D> *positions, int N, Scalar t, Scalar one_minus_t)
 {
 	for (int bezier_order = 1; bezier_order < N; bezier_order++) {
